requirements.txt
dash
dash_leaflet
requests
beautifulsoup4
selenium
selectolax
chompjs
parsel
requests_html
lxml_html_clean
pymongo
python-dotenv